Refactor: getProfileLikes customHook으로 분리
Related to: #321
