+++ src/types/likeType.ts
@@ -1,0 +1,9 @@
+export interface LikedChallengeDataType {
+  instanceId: number;
+  title: string;
+  pointPerPerson: number;
+  fileResponse: {
+    encodedFile: string;
+  };
+  likesId: number;
+}
